test(ProductAcceptanceTest): Product 인수테스트 작성
- Product 생성 팩토리메소드 수정

+++ src/main/java/kitchenpos/domain/Product.java
@@ -21,8 +21,8 @@ public class Product {
         this.price = price;
     }
 
-    public static Product of(String name, BigDecimal price) {
-        return new Product(name, price);
+    public static Product of(String name, Integer price) {
+        return new Product(name, BigDecimal.valueOf(price));
     }
 
     public Long getId() {
